Fix admin page spacing issue.
